Release 5
additions
+ new Fast Scaffold mode,s Jump B, Jump C
+ added precision slider to Scaffold
+ all mods besides forge defaults are now hidden from being sent to server
+ added Prioritize enemies option to KillAura
+ added third person distance to NoCameraClip

fixes, removals, misc
+ improved Scaffold block placements
+ improved Scaffold rotations
+ fixed enemy and friend commands in Command Line
+ removed Fix slot reset due to its uselessness
+ fixed players not being removed as enemies when friended
+ improved nametags (fixed enchant glint, fixed blocks rotating, improved autoscale)
+ improved scripting API
+ improved Vanilla tower in Tower

+++ src/main/java/keystrokesmod/mixins/impl/render/MixinEntityRenderer.java
@@ -75,8 +75,9 @@ public class MixinEntityRenderer {
                     GlStateManager.rotate(entity.prevRotationYaw + (entity.rotationYaw - entity.prevRotationYaw) * partialTicks + 180.0F, 0.0F, -1.0F, 0.0F);
                     GlStateManager.rotate(entity.prevRotationPitch + (entity.rotationPitch - entity.prevRotationPitch) * partialTicks, -1.0F, 0.0F, 0.0F);
                 }
-            } else if (this.mc.gameSettings.thirdPersonView > 0) {
-                double d3 = (double) (this.thirdPersonDistanceTemp + (this.thirdPersonDistance - this.thirdPersonDistanceTemp) * partialTicks);
+            }
+            else if (this.mc.gameSettings.thirdPersonView > 0) {
+                double d3 = this.thirdPersonDistanceTemp + (this.thirdPersonDistance - this.thirdPersonDistanceTemp) * partialTicks;
                 if (this.mc.gameSettings.debugCamEnable) {
                     GlStateManager.translate(0.0F, 0.0F, (float) (-d3));
                 } else {
